+++ Runtime/AssemblyInfo.cs
@@ -1,0 +1,6 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo(
+    "Deucarian.ViewerNavigation.Tests.EditMode")]
+[assembly: InternalsVisibleTo(
+    "Deucarian.ViewerNavigation.Tests.PlayMode")]